requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==11.0.0
numpy==2.1.0
imagehash==4.3.1
requests==2.31.0
gunicorn==21.2.0
